% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_paper.R
\name{add_csl_environment}
\alias{add_csl_environment}
\title{Add CSL Environment}
\usage{
add_csl_environment(x)
}
\arguments{
\item{x}{Character scalar with the intermediate markdown file}
}
\value{
Character scalar with the intermediate markdown file with changes made.
}
\description{
Replaces the temporary place holders for the CSL environment with Latex-formatted
begin and end designations.
}
