% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble-graphs.R
\name{build_label_grob}
\alias{build_label_grob}
\title{Build Label Grob}
\usage{
build_label_grob(label_type, plot, w = unit(6.5, "in"))
}
\arguments{
\item{label_type}{Character scalar giving the label to convert to a grob.}

\item{plot}{A ggplot graph}

\item{w}{The width of the plot expressed as a grid unit. Default is `unit(6.5, 'in')`.}
}
\value{
A textbox_grob from the gridtext package.
}
\description{
Extracts the necessary information from a ggplot object to convert one of the
labels (title, subtitle, caption, or tag) into a textbox_grob.
}
