% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-notes.R
\name{calc_grob_width}
\alias{calc_grob_width}
\title{Caculate Width of Graphical Object}
\usage{
calc_grob_width(gr, unit)
}
\arguments{
\item{gr}{A graphical object}

\item{unit}{Character scalar providing the unit to use in calculating the width
(i.e., 'in', 'mm')}
}
\description{
Returns the width of a graphical object (grob) in the specifie units.
}
