% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-label.R
\name{compute_edgelabel}
\alias{compute_edgelabel}
\title{Compute Edge Label}
\usage{
compute_edgelabel(self, data, params, layout)
}
\arguments{
\item{self}{A ggproto object}

\item{data}{The data object supplied to the ggplot object}

\item{params}{List of parameters supplied to the stat_edgelabel
function}

\item{layout}{Object providing infomration about the plot layout,
including the range of th x and y axes}
}
\value{
A data.frame with the processed data.
}
\description{
This is the computational engine for stat_edgelabel.  It takes
the available data a grabs the right-most values of each group,
where it places a label.
}
