% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-notes.R
\name{create_noteGrob}
\alias{create_noteGrob}
\title{Create a Note Graphical Object}
\usage{
create_noteGrob(note_text, gp, as_textGrob = FALSE)
}
\arguments{
\item{note_text}{Character scalar of text to put in the note}

\item{gp}{Graphic parameter list covering how the text is to be formatted.
Created by call to grid::gpar.}
}
\description{
Creates the noteGrob object
}
