% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-label.R
\name{geom_edgelabel}
\alias{geom_edgelabel}
\title{geom_edgelabel}
\usage{
geom_edgelabel(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  hjust = 0,
  size = 10,
  inherit.aes = TRUE,
  nudge_y = 0,
  nudge_x = 0,
  ...
)
}
\arguments{
\item{nudge_y}{Named character vector to use to nudge the y-values.  Nudges
should be expressed in 'npc' units, which capture the percentage of the
plot area. So setting nudge_y = 2L will move each label up 2% of the viewable
plot area.}

\item{nudge_x}{The number of characters to move each label to the right.
A value of "1L" adds a single empty space to each label.  A value of 2L
adds two spaces.  Default is 0.}
}
\description{
Adds labels to the right of the plot for each group in the data. The
labels appear to the right of the farthest right point of any of the
plotted series.
}
