% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-recession.R
\name{geom_recession}
\alias{geom_recession}
\title{NBER Recession Date Geom}
\usage{
geom_recession(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  hjust = 0,
  size = 10,
  inherit.aes = TRUE,
  nudge_y = 0,
  nudge_x = 0,
  alpha = 0.1,
  fill = "gray90",
  ...
)
}
\description{
Adds shaded areas to a time series that indicate the periods corresponding
to recessions as dated by the National Bureau of Economic Research.
}
