% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_paper.R
\name{get_from_parent_frame}
\alias{get_from_parent_frame}
\title{Get from Parent Frame}
\usage{
get_from_parent_frame(x)
}
\arguments{
\item{x}{Scalar character vector giving the name of the variable to be searched
for.}
}
\value{
The value of the variable, if found. If not found returns NULL.
}
\description{
Searches parent frames to find a specified variable.
}
