% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-notes.R
\name{identify_best_break}
\alias{identify_best_break}
\title{Identify Best Break}
\usage{
identify_best_break(txt, width, unit, gp)
}
\arguments{
\item{txt}{Character scalar of note text}

\item{width}{Numeric scalar giving width of available space}

\item{unit}{Character scalar giving the units of measurement to be used.}

\item{gp}{Graphical parameters to be used in printing the text. Generated by
grid::gpar.}
}
\description{
Finds the position in the character scalar that contains the most characters
without exceeding the specified width.
}
