% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-label.R
\name{nudge_y_labels}
\alias{nudge_y_labels}
\title{Nudge Y Labels}
\usage{
nudge_y_labels(label, nudge_y)
}
\arguments{
\item{label}{Character vector of labels}

\item{nudge_y}{Named character vector supplied by the  user.  The name
of each element will correspond to the label to which the nudge is to
be applied.}
}
\value{
A numeric vector of nudges
}
\description{
Creates a vector of nudges to apply to the y-axis of the label.
}
