% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-label.R
\name{pad_label}
\alias{pad_label}
\title{Pad Label}
\usage{
pad_label(x, s = 1L)
}
\arguments{
\item{x}{A vector of character labels to be displayed}

\item{s}{Integer scalar giving the number of blank spaces to use in
nudging the label to the right.}
}
\value{
A character vector with the extra blank spaces added
}
\description{
Adds additional empty spaces before the labels being displayed along
the edge. This allows the x position of each label to remain at the
same point as the data to avoid the labels being completely dropped
by the limit set for the x-axis.
}
