% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_paper.R
\name{replace_duplicate_punctuation}
\alias{replace_duplicate_punctuation}
\title{Replace Duplicate Punctuation}
\usage{
replace_duplicate_punctuation(x)
}
\arguments{
\item{x}{Character scalar that contains all the text in the intermediate markdown file}
}
\value{
Character scalar with the duplicate punctuation removed
}
\description{
In some cases, CSL bibliographies may allow redundant punctuation.  For example,
when an item's title ends in a question mark, this will frequently result in
the title ending in a "?." which looks ugly.  This function searchs the bibliography
for such duplicates and, when found, removes the second punctuation.
}
