% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble-graphs.R
\name{save_kpb_graph}
\alias{save_kpb_graph}
\title{Save KPB Graph}
\usage{
save_kpb_graph(
  filename,
  plot = last_plot(),
  device = NULL,
  width = 6.5,
  units = "in"
)
}
\arguments{
\item{filename}{Character vector giving the output file path}

\item{plot}{Ggplot graph to save.  Default's to the last plot.}

\item{device}{A graphical device to use to print to.  Default is
NULL in which case the device will be assumed based on the extension
given the output file.}

\item{width}{Numerical scalar giving the width of the output. Default is
6.5.}

\item{units}{Character scalar giving the units in which the width is
provided.  Default equals 'in'.}
}
\description{
Converts a ggplot2 graphic into an output format that files the style
I like to use.
}
