% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-label.R
\name{stat_edgelabel}
\alias{stat_edgelabel}
\title{Stat Edge Label}
\usage{
stat_edgelabel(
  data = NULL,
  mapping = NULL,
  position = "identity",
  ...,
  inherit.aes = TRUE
)
}
\description{
A `ggplot2` stat that takes the labels for series of data and
labels each series at its right-most point.
}
