% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_print}
\alias{theme_print}
\title{Presentation Theme}
\usage{
theme_print(
  legend.position = "none",
  as.ts = FALSE,
  panel.grid.major.x = element_blank(),
  panel.grid.minor.x = element_blank(),
  panel.grid.major.y = element_blank(),
  panel.grid.minor.y = element_blank()
)
}
\description{
This is a ggplot2 theme that I have prepared to format graphs in my
preferred style for presentations, which uses larger fonts than would
be appropriate in a printed graph.
}
