% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_paper.R
\name{use_default_csl}
\alias{use_default_csl}
\title{Use Default CSL}
\usage{
use_default_csl()
}
\value{
A character scalar giving the full path to brevoort.csl
}
\description{
A convenience function that allows users to use a default CSL file
that is included with the package. The default is brevoort.csl, which
is a custom CSL specification that I created.
}
