% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-notes.R
\name{wrap_notes}
\alias{wrap_notes}
\title{Wrap Notes}
\usage{
wrap_notes(text, width = 3, unit = "in", gp = set_gp())
}
\arguments{
\item{text}{Character scalar with the text to be included.}

\item{width}{Numeric scalar giving the alloted width for the viewport}

\item{unit}{Character scalar giving the units in which the width is expressed
(i.e., 'in', 'mm').}
}
\description{
Takes a character string that is to appear as notes below a graph and places
line breaks at appropriate places to wrap the line to conform to the appropriate
width biven the supplied font size.
}
