% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_presentation}
\alias{theme_presentation}
\title{Presentation Theme}
\usage{
theme_presentation(legend.position = "none")
}
\description{
This is a ggplot2 theme that I have prepared to format graphs in my
preferred style for presentations, which uses larger fonts than would
be appropriate in a printed graph.
}
